A function for both working API
The translateAPI.py file contains a function in which you present the path of the image on the parameter and it returns you a dictionary with keywords of the identified items with its own score of how accurate the keyword matches the picture analyzed.  The googleAPI.py file contains a function in which you put a dictionary variable and the language you want to change the current dictionary to and it returns a list of keywords in the language desired to translate to.  The apiTutorial-58994ef72fa.json file contains the permission to acces the google translate API.